+++ internal/provider/helpers/utils.go
@@ -42,3 +42,19 @@ func GetValueSlice(result []gjson.Result) []attr.Value {
 	}
 	return v
 }
+
+func GetStringList(result []gjson.Result) types.List {
+	v := make([]attr.Value, len(result))
+	for r := range result {
+		v[r] = types.StringValue(result[r].String())
+	}
+	return types.ListValueMust(types.StringType, v)
+}
+
+func GetInt64List(result []gjson.Result) types.List {
+	v := make([]attr.Value, len(result))
+	for r := range result {
+		v[r] = types.Int64Value(result[r].Int())
+	}
+	return types.ListValueMust(types.Int64Type, v)
+}