+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.1.14 (unreleased)
 
 - Add `iosxe_prefix_list` resource and data source
+- Add `iosxe_route_map` resource and data source
 
 ## 0.1.13